// src/pages/History/types.ts
export interface Medical {
  name: '';
  date: '';
  activeFlag: true;
}
export interface Surgical {
  name: '';
  date: '';
}
export interface Social {
  smokingHistory: '';
  smquantity: '';
  smtime: '';
  tobacco: '';
  tValue: '';
  alcohol: '';
  alStatus: '';
  exercise: '';
  exerciseValues: '';
  marijuana: '';
}